web-animations/animation-player/animation-player-current-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationPlayer currentTime attribute tests</title>
<meta name="assert" content="The current time of this player unless this player has a pending pause task, in which case this attribute returns null">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationplayer-currenttime">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
var timesToTest = [-100,
    -1,
    0,
    1,
    100,
    ANIMATION_END_TIME / 2,
    ANIMATION_END_TIME - 100,
    ANIMATION_END_TIME,
    ANIMATION_END_TIME + 1,
    ANIMATION_END_TIME + 100];

function calculateCurrentTime(timelineTime, startTime, playbackRate) {
    return (timelineTime - startTime) * playbackRate;
}

function calculateStartTime(timelineTime, seekTime, playbackRate) {
    return timelineTime - seekTime / playbackRate;
}

test(function() {
    var player = new AnimationPlayer();
    assert_equals(player.currentTime, null, 'AnimationPlayer.currentTime should be null');
}, 'Test that AnimationPlayer.currentTime attribute is null if player has no associated timeline');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));

    player.ready.then(t.step_func(function() {
        assert_equals(player.currentTime, calculateCurrentTime(
            player.timeline.currentTime, player.startTime, player.playbackRate),
            'AnimationPlayer.currentTime value should be calculated as defined');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.currentTime value is (timeline time - start time) × playback rate');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = 0.4;

    player.ready.then(t.step_func(function() {
        setTimeout(function() {
            assert_equals(player.currentTime, calculateCurrentTime(
                player.timeline.currentTime, player.startTime, player.playbackRate),
                'AnimationPlayer.currentTime value should be calculated as defined');
            t.done();
        }, 50);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.currentTime value is (timeline time - start time) × playback rate. ' +
    'Test 0 < playbackRate < 1');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = 2;

    player.ready.then(t.step_func(function() {
        setTimeout(function() {
            assert_equals(player.currentTime, calculateCurrentTime(
                player.timeline.currentTime, player.startTime, player.playbackRate),
                'AnimationPlayer.currentTime value should be calculated as defined');
            t.done();
        }, 50);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.currentTime value is (timeline time - start time) × playback rate. ' +
    'Test playbackRate > 1');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = -1.5;

    player.ready.then(t.step_func(function() {
        player.currentTime = ANIMATION_END_TIME / 2;
        setTimeout(function() {
            assert_equals(player.currentTime, calculateCurrentTime(
                player.timeline.currentTime, player.startTime, player.playbackRate),
                'AnimationPlayer.currentTime value should be calculated as defined');
            t.done();
        }, 50);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.currentTime value is (timeline time - start time) × playback rate. ' +
    'Test playbackRate < 0');

// 15.10.2014 Finished here. Go on from this place

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.startTime = 100;
    setTimeout(function() {
        t.step(function () {
            assert_equals(player.currentTime, calculateCurrentTime(
                player.timeline.currentTime, player.startTime, player.playbackRate),
                'AnimationPlayer.currentTime value should be ' +
                'effective current time of this player');
            t.done();
        });
    }, 100);
}, 'AnimationPlayer.currentTime value is effective current time of this player. ' +
    'Start time > 0 and wait for some time before checking the value');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = -1;
    player.startTime = 500;
    setTimeout(function() {
        t.step(function () {
            assert_equals(player.currentTime, calculateCurrentTime(
                player.timeline.currentTime, player.startTime, player.playbackRate),
                'AnimationPlayer.currentTime value should be ' +
                'effective current time of this player');
            t.done();
        });
    }, 100);
}, 'AnimationPlayer.currentTime value is effective current time of this player. ' +
    'Wait for some time is before checking the value, playbackRate < 0');

test(function() {
    var player = newAnimationPlayer(createDiv(this));

    timesToTest.forEach(function(val) {
        player.currentTime = val;
        assert_equals(player.currentTime, val,
            'AnimationPlayer.currentTime value should be set to the new value');
    });
}, 'Test setting value of AnimationPlayer.currentTime attribute');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.pause();

    timesToTest.forEach(function(val) {
        player.currentTime = val;
        assert_equals(player.currentTime, val,
            'AnimationPlayer.currentTime value should be set to the new value');
    });
}, 'Test setting value of AnimationPlayer.currentTime attribute when player is paused');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = 0;

    timesToTest.forEach(function(val) {
        player.currentTime = val;
        assert_equals(player.currentTime, val,
            'AnimationPlayer.currentTime value should be set to the new value');
    });
}, 'Test setting value of AnimationPlayer.currentTime attribute ' +
    'when player playback rate is zero');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = 100;
    player.playbackRate = 0;

    timesToTest.forEach(function(val) {
        player.currentTime = val;
        assert_equals(player.currentTime, val,
            'AnimationPlayer.currentTime value should be set to the new value');
    });
}, 'Test setting value of AnimationPlayer.currentTime attribute ' +
    'when player playback rate is zero and seek performed before setting playbackRate');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.pause();
    player.playbackRate = 0;

    timesToTest.forEach(function(val) {
        player.currentTime = val;
        assert_equals(player.currentTime, val,
            'AnimationPlayer.currentTime value should be set to the new value');
    });
}, 'Test setting value of AnimationPlayer.currentTime attribute ' +
    'when player playback rate is zero and player is paused');

test(function() {
    var player = newAnimationPlayer(createDiv(this));

    timesToTest.forEach(function(val) {
        player.playbackRate = 0;
        player.currentTime = val;
        assert_equals(player.currentTime, val,
            'AnimationPlayer.currentTime value should be set to the new value');

        player.playbackRate = 1;
        assert_equals(player.currentTime, val, 'AnimationPlayer.currentTime value ' +
            'should not be changed after changing playbackRate');
    });
}, 'Test changing of AnimationPlayer.currentTime value ' +
    'when player playback rate is changed from zero to 1 and back');

test(function() {
    var player = newAnimationPlayer(createDiv(this));

    timesToTest.forEach(function(seekTime) {
        var expectedTime = calculateStartTime(player.timeline.currentTime, seekTime,
            player.playbackRate);
        player.currentTime = seekTime;

        assert_equals(player.startTime, expectedTime,
            'The value of AnimationPlayer.startTime attribute should be set when seek ' +
            'is performed');
        assert_false(player.finished, 'Player\'s finished flag should be set to false');
    });
}, 'Test that when player is not paused and playback rate is not zero ' +
    'setting value of AnimationPlayer.currentTime attribute sets start time and finished flag');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = 0.4;

    timesToTest.forEach(function(seekTime) {
        var expectedTime = calculateStartTime(player.timeline.currentTime, seekTime,
            player.playbackRate);
        player.currentTime = seekTime;

        assert_equals(player.startTime, expectedTime,
            'The value of AnimationPlayer.startTime attribute should be set when seek ' +
            'is performed');
        assert_false(player.finished, 'Player\'s finished flag should be set to false');
    });
}, 'Test that when player is not paused and playback rate is not zero setting ' +
    'value of AnimationPlayer.currentTime attribute sets start time and finished flag.' +
    'Playback rate is > 0 and < 1');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = -0.4;

    seekTimes.forEach(function(seekTime) {
        var expectedTime = calculateStartTime(player.timeline.currentTime, seekTime,
            player.playbackRate);
        player.currentTime = seekTime;

        assert_equals(player.startTime, expectedTime,
            'The value of AnimationPlayer.startTime attribute should be set when seek ' +
            'is performed');
        assert_false(player.finished, 'Player\'s finished flag should be set to false');
    });
}, 'Test that when player is not paused and playback rate is not zero setting ' +
    'value of AnimationPlayer.currentTime attribute sets start time and finished flag.' +
    'Playback rate is < 0 and > -1');

test(function() {
    var animation = newAnimation(createDiv(this));
    var player = document.timeline.play(animation);

    assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_0,
        'Animation effect should take place');
}, 'Test that animation effect is changed accordingly by setting ' +
    'AnimationPlayer.currentTime attribute when current time is zero');

test(function() {
    var animation = newAnimation(createDiv(this));
    var player = document.timeline.play(animation);
    player.currentTime = ANIMATION_END_TIME / 2;

    assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_0_5,
        'Animation effect should take place');
}, 'Test that animation effect is changed accordingly by setting ' +
    'AnimationPlayer.currentTime attribute when current time is > 0 and < end time');

test(function() {
    var animation = newAnimation(createDiv(this));
    var player = document.timeline.play(animation);
    player.currentTime = ANIMATION_END_TIME;

    assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_1,
        'Animation effect should take place');
}, 'Test that animation effect is changed accordingly by setting ' +
    'AnimationPlayer.currentTime attribute when current time equals end time');
</script>
</body>
